(ref/services): refactor services and test cases fixes in application.properties and application-test.properties
