+++ pom.xml
@@ -29,6 +29,7 @@
         <wiremock.version>3.12.0</wiremock.version>
         <sonar-maven-plugin.version>3.8.0.2131</sonar-maven-plugin.version>
         <spring-cloud-dependencies.version>2022.0.4</spring-cloud-dependencies.version>
+        <springdoc-openapi-starter-webmvc-ui.version>2.6.0</springdoc-openapi-starter-webmvc-ui.version>
         <maven-dependency-plugin.version>3.8.1</maven-dependency-plugin.version>
     </properties>
     <dependencies>
@@ -103,7 +104,7 @@
         <dependency>
             <groupId>org.springdoc</groupId>
             <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
-            <version>2.6.0</version>
+            <version>${springdoc-openapi-starter-webmvc-ui.version}</version>
         </dependency>
 
 <!--        test -->
