+++ src/main/java/dev/shann/mcproductservice/utils/ApplicationConstants.java
@@ -1,0 +1,6 @@
+package dev.shann.mcproductservice.utils;
+
+public final class ApplicationConstants {
+    public static final String INVALID_USER = "Invalid User";
+    public static final String AUTHENTICATE = "/authenticate";
+}
