+++ src/main/resources/application.properties
@@ -11,9 +11,6 @@ spring.jpa.database-platform= org.hibernate.dialect.MySQLDialect
 spring.jpa.hibernate.ddl-auto= update
 #default value for pagination = 2000
 spring.data.jpa.properties.max-results-per-page= 4
-#spring.datasource.url=jdbc:h2:mem: mcproductservice
-#spring.datasource.driver-class-name= org.h2.Driver
-#spring.jpa.database-platform= org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql= true
 #log setup
 logging.level.org.hibernate.SQL=DEBUG
@@ -21,8 +18,8 @@ logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 logging.level.org.hibernate.orm.jdbc.bind=TRACE
 logging.file.name = log/product-service.log
 #spring.main.allow-bean-definition-overriding= true
-user.service.client.user.name: http://user-service/users
-user.service.client.user.name-alt: http://localhost:8082/users
+user.service.client.user.name= http://user-service/userService/api/v1/users
+user.service.client.user.name-alt= http://localhost:8082/userService/api/v1/users
 # actuator
 management.endpoints.web.exposure.include=health, info, loggers, metrics
 # spring mail
