+++ src/test/resources/application-test.properties
@@ -4,7 +4,7 @@ spring.datasource.username= root
 spring.datasource.password= root
 spring.jpa.database-platform= org.hibernate.dialect.H2Dialect
 spring.jpa.show-sql=true
-user.service.client.user.name-alt: http://localhost:8082/users
+user.service.client.user.name-alt= http://localhost:8082/users
 # spring mail
 spring.mail.host=localhost
 spring.mail.port=1025
